move wagmi files into one folder
