--- packages/arb-token-bridge-ui/src/util/arbitrumNova.ts
@@ -1,21 +1,0 @@
-import { Chain } from 'wagmi'
-import { ChainId, rpcURLs } from './networks'
-
-export const arbitrumNova: Chain = {
-  id: ChainId.ArbitrumNova,
-  name: 'Arbitrum Nova',
-  network: 'arbitrum-nova',
-  nativeCurrency: { name: 'Ether', symbol: 'ETH', decimals: 18 },
-  rpcUrls: {
-    default: {
-      http: [rpcURLs[ChainId.ArbitrumNova]!]
-    },
-    public: {
-      http: [rpcURLs[ChainId.ArbitrumNova]!]
-    }
-  },
-  blockExplorers: {
-    etherscan: { name: 'Arbiscan', url: 'https://nova.arbiscan.io' },
-    default: { name: 'Arbiscan', url: 'https://nova.arbiscan.io' }
-  }
-}
