+++ packages/arb-token-bridge-ui/src/util/wagmi/wagmiAdditionalNetworks.ts
@@ -1,6 +1,28 @@
 import { Chain } from 'wagmi'
-import { ChainId, rpcURLs } from './networks'
+import { ChainId, rpcURLs } from '../networks'
 
+export const arbitrumNova: Chain = {
+  id: ChainId.ArbitrumNova,
+  name: 'Arbitrum Nova',
+  network: 'arbitrum-nova',
+  nativeCurrency: { name: 'Ether', symbol: 'ETH', decimals: 18 },
+  rpcUrls: {
+    default: {
+      http: [rpcURLs[ChainId.ArbitrumNova]!]
+    },
+    public: {
+      http: [rpcURLs[ChainId.ArbitrumNova]!]
+    }
+  },
+  blockExplorers: {
+    etherscan: { name: 'Arbiscan', url: 'https://nova.arbiscan.io' },
+    default: { name: 'Arbiscan', url: 'https://nova.arbiscan.io' }
+  }
+}
+
+/**
+ * For e2e testing
+ */
 export const localL1Network: Chain = {
   id: ChainId.Local,
   name: 'EthLocal',
@@ -19,6 +41,9 @@ export const localL1Network: Chain = {
   }
 }
 
+/**
+ * For e2e testing
+ */
 export const localL2Network: Chain = {
   id: ChainId.ArbitrumLocal,
   name: 'ArbLocal',
